pyproject.toml
[tool.poetry]
name = "webhook"
version = "0.1.0"
description = ""
authors = ["José <josepsousa2012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.1"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.21.5"
celery = "^5.2.7"
psycopg = "^3.1.9"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
